requirements.txt
Flask
requests
pandas
numpy
polyline
matplotlib
seaborn
scikit-learn
python-dotenv